fix(process): one captureProcessStartTime, with the Windows branch
Review caught that the win32 branch landed in only ONE of the two copies.
teams/agents.ts carried its own near-identical implementation with no Windows
support, so 'agents teams stop' still had no pid-reuse protection there -- the
exact failure its own docstring warns about.

Hoists a single implementation into platform/process.ts; pty-server.ts and
teams/agents.ts both delegate. Three fixes on the way through: ToFileTimeUtc()
so the fingerprint is culture-independent (the bare DateTime stringifies under
the current locale), a 5s timeout so a wedged WMI service cannot hang the PTY
sidecar, windowsHide plus Number.isInteger(pid) to match the repo's other
console spawns, and memoization per pid -- the sessions listing probes one pid
per row, and a PowerShell cold start each was not acceptable there.
